requirements.txt
# Core dependencies for Mutual Fund FAQ Assistant

# Web scraping and HTTP requests
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
lxml>=4.9.0

# PDF processing
PyPDF2>=3.0.1
pdfplumber>=0.9.0

# Data processing and analysis
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Text processing and NLP
nltk>=3.8.0
spacy>=3.6.0
textblob>=0.17.0

# Vector databases and embeddings
pinecone>=2.2.0
weaviate-client>=3.24.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
transformers>=4.33.0
torch>=2.0.0
FlagEmbedding>=1.2.0

# LLM APIs
openai>=1.0.0
groq>=0.4.0

# Task queue and background processing
celery>=5.3.0
redis>=4.6.0

# Web framework
fastapi>=0.103.0
uvicorn>=0.23.0
jinja2>=3.1.0

# Configuration management
pydantic>=2.3.0
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Logging and monitoring
structlog>=23.1.0
prometheus-client>=0.17.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.24.0

# Development tools
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0
pre-commit>=3.4.0

# Security
cryptography>=41.0.0
python-jose>=3.3.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3
